440: move Edge, Graph, and Point to global/class
