+++ src/global/class/Edge.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import { assert } from '../../global/utils';
 import { Point } from './Point';
-import { detD, threePointsInLine } from '../utils';
+import { assert } from '../utils';
+import { detD, threePointsInLine } from '../../interpret/utils';
 
 
 export class Edge {
