+++ src/global/class/Point.js
