+++ src/plan/test/backwardsFunnelData.js
@@ -1,4 +1,4 @@
-import { Point } from '../../interpret/class/Point';
+import { Point } from '../../global/class/Point';
 import { Polypoint } from '../../interpret/class/Polypoint';
 import { Triangle } from '../../interpret/class/Triangle';
 import { PolypointState } from '../class/PolypointState';
